+++ packages/categorize/CHANGELOG.md
@@ -3,6 +3,122 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.2.3...@pie-element/categorize@4.3.0) (2019-06-21)
+
+
+### Features
+
+* **general:** updated all the packages that have the config-ui as a dependency ([44061de](https://github.com/pie-framework/pie-elements/commit/44061de))
+
+
+
+
+
+## [4.2.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.2.2...@pie-element/categorize@4.2.3) (2019-06-20)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.2.1...@pie-element/categorize@4.2.2) (2019-06-20)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.2.0...@pie-element/categorize@4.2.1) (2019-06-19)
+
+
+### Bug Fixes
+
+* **scoring-and-feedback:** fixed scoring and added alternate responses ([dd0bb02](https://github.com/pie-framework/pie-elements/commit/dd0bb02))
+
+
+
+
+
+# [4.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.1.2...@pie-element/categorize@4.2.0) (2019-06-19)
+
+
+### Features
+
+* **extras:** added row labels support and alternate responses ([280028f](https://github.com/pie-framework/pie-elements/commit/280028f))
+
+
+
+
+
+## [4.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.1.1...@pie-element/categorize@4.1.2) (2019-06-18)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.1.0...@pie-element/categorize@4.1.1) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+# [4.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.6...@pie-element/categorize@4.1.0) (2019-06-13)
+
+
+### Features
+
+* **ui:** updated the ui ([2baa579](https://github.com/pie-framework/pie-elements/commit/2baa579))
+
+
+
+
+
+## [4.0.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.5...@pie-element/categorize@4.0.6) (2019-06-12)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.0.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.4...@pie-element/categorize@4.0.5) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.0.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.3...@pie-element/categorize@4.0.4) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.0.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.2...@pie-element/categorize@4.0.3) (2019-06-05)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
+## [4.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.1...@pie-element/categorize@4.0.2) (2019-05-23)
+
+**Note:** Version bump only for package @pie-element/categorize
+
+
+
+
+
 ## [4.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/categorize@4.0.0...@pie-element/categorize@4.0.1) (2019-05-21)