+++ packages/drawing-response/CHANGELOG.md
@@ -2,3 +2,42 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [1.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response@1.0.4...@pie-element/drawing-response@1.1.0) (2019-06-21)
+
+
+### Features
+
+* **general:** updated all the packages that have the config-ui as a dependency ([44061de](https://github.com/pie-framework/pie-elements/commit/44061de))
+
+
+
+
+
+## [1.0.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response@1.0.3...@pie-element/drawing-response@1.0.4) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/drawing-response
+
+
+
+
+
+## [1.0.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response@1.0.2...@pie-element/drawing-response@1.0.3) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/drawing-response
+
+
+
+
+
+## [1.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/drawing-response@1.0.1...@pie-element/drawing-response@1.0.2) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/drawing-response
+
+
+
+
+
+## 1.0.1 (2019-05-31)
+
+**Note:** Version bump only for package @pie-element/drawing-response