+++ packages/graph-lines/configure/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.7...@pie-element/graph-lines-configure@2.2.0) (2019-06-21)
+
+
+### Features
+
+* **general:** updated all the packages that have the config-ui as a dependency ([44061de](https://github.com/pie-framework/pie-elements/commit/44061de))
+
+
+
+
+
+## [2.1.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.6...@pie-element/graph-lines-configure@2.1.7) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
+## [2.1.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.5...@pie-element/graph-lines-configure@2.1.6) (2019-06-12)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
+## [2.1.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.4...@pie-element/graph-lines-configure@2.1.5) (2019-06-07)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
+## [2.1.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.3...@pie-element/graph-lines-configure@2.1.4) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
+## [2.1.3](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.2...@pie-element/graph-lines-configure@2.1.3) (2019-06-05)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
+## [2.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.1...@pie-element/graph-lines-configure@2.1.2) (2019-05-23)
+
+**Note:** Version bump only for package @pie-element/graph-lines-configure
+
+
+
+
+
 ## [2.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/graph-lines-configure@2.1.0...@pie-element/graph-lines-configure@2.1.1) (2019-05-21)
 
 **Note:** Version bump only for package @pie-element/graph-lines-configure