+++ packages/inline-dropdown/controller/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.4.1...@pie-element/inline-dropdown-controller@1.4.2) (2019-06-17)
+
+**Note:** Version bump only for package @pie-element/inline-dropdown-controller
+
+
+
+
+
+## [1.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.4.0...@pie-element/inline-dropdown-controller@1.4.1) (2019-06-13)
+
+**Note:** Version bump only for package @pie-element/inline-dropdown-controller
+
+
+
+
+
+# [1.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.3.0...@pie-element/inline-dropdown-controller@1.4.0) (2019-06-12)
+
+
+### Features
+
+* **correct-responses:** updated elements with showing correct responses when needed ([8ea0d12](https://github.com/pie-framework/pie-elements/commit/8ea0d12))
+
+
+
+
+
+# [1.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.2.1...@pie-element/inline-dropdown-controller@1.3.0) (2019-06-07)
+
+
+### Features
+
+* **mask-markup-elements:** enabled partial scoring for the 3 elements and added a controller to the match-list ([41e210e](https://github.com/pie-framework/pie-elements/commit/41e210e))
+
+
+
+
+
+## [1.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.2.0...@pie-element/inline-dropdown-controller@1.2.1) (2019-06-06)
+
+**Note:** Version bump only for package @pie-element/inline-dropdown-controller
+
+
+
+
+
+# [1.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.1.2...@pie-element/inline-dropdown-controller@1.2.0) (2019-06-05)
+
+
+### Features
+
+* **mask-markup-elements:** major changes to all 3 elements ([039a33e](https://github.com/pie-framework/pie-elements/commit/039a33e))
+* **mask-markup-elements:** most of the work for the 3 new elements: inline-dropdown, explicit-constructed-response and drag-in-the-blank ([67f3c3d](https://github.com/pie-framework/pie-elements/commit/67f3c3d))
+
+
+
+
+
 ## [1.1.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown-controller@1.1.1...@pie-element/inline-dropdown-controller@1.1.2) (2019-05-08)
 
 **Note:** Version bump only for package @pie-element/inline-dropdown-controller